backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-dotenv
cryptography
requests
pyodbc
mysql.connector

python-jose[cryptography]
